refactor(font_factory): Internal implementation of font factory. Moved the implementation of the font factory to the internal graphics namespace. The moth font factory now just takes the internal type as a dependency.
