+++ include/graphics/moth_ui/moth_font_factory.h
@@ -1,20 +1,19 @@
 #pragma once
 
-#include "graphics/igraphics.h"
+#include "graphics/font_factory.h"
 #include <moth_ui/font_factory.h>
 
 namespace graphics {
     class MothFontFactory : public moth_ui::FontFactory {
     public:
-        MothFontFactory(graphics::IGraphics& graphics);
+        MothFontFactory(std::shared_ptr<graphics::FontFactory> factoryImpl);
         virtual ~MothFontFactory() = default;
 
         void ClearFonts() override;
         std::shared_ptr<moth_ui::IFont> GetFont(char const* name, int size) override;
 
     private:
-        graphics::IGraphics& m_graphics;
-        std::map<std::string, std::map<int, std::shared_ptr<moth_ui::IFont>>> m_fonts;
+        std::shared_ptr<graphics::FontFactory> m_factoryImpl;
     };
 }
